Extract inference target resolution into inference_target module.
Unify accelerated backfill segment loading for runtime inference and corpus coverage; stop importing resolution from analytic.py.
